Fixed opt order

+++ src/search/symbolic/opt_order.h
@@ -6,16 +6,16 @@
 
 namespace symbolic {
 class InfluenceGraph {
-    std::vector<std::vector<long>> influence_graph;
+    std::vector<std::vector<double>> influence_graph;
 
-    long influence(int v1, int v2) const {
+    double influence(int v1, int v2) const {
         return influence_graph[v1] [v2];
     }
 
 
-    long optimize_variable_ordering_gamer(std::vector <int> &order,
+    double optimize_variable_ordering_gamer(std::vector <int> &order,
                                           int iterations) const;
-    long compute_function(const std::vector <int> &order) const;
+    double compute_function(const std::vector <int> &order) const;
     void optimize_ordering_gamer(std::vector <int> &ordering) const;
     static void randomize(std::vector <int> &ordering, std::vector<int> &new_order);
 
@@ -28,7 +28,7 @@ public:
                                           std::vector <int> &partition_sizes,
                                           int iterations = 50000) const;
 
-    void set_influence(int v1, int v2, long val = 1) {
+    void set_influence(int v1, int v2, double val = 1) {
         influence_graph[v1][v2] = val;
         influence_graph[v2][v1] = val;
     }
